Changed the code to act much more like a hierarchal model.
Actually connect the various components into a full model, now, as opposed to just repeating the code to create them. Added a hierchal model type that can be traversed like a tree, allowing arbitrary closed model solids that are representable by a set of vectors to be added to a model.